requirements.txt
discord.py = 1.7.3
requests
urllib
python-decouple
ffmpeg = 1.4
ffprobe = 0.5
PyNaCl = 1.4.0
yt-dlp = 2021.10.10